rename sinenet_v1
